backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
gunicorn==23.0.0
motor==3.6.0
redis[asyncio]==5.2.0
pydantic==2.9.0
pydantic-settings==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
groq==0.12.0
boto3>=1.35.0
python-multipart==0.0.12
fastapi-mail==1.4.1
httpx==0.27.0
